rename(public-data): 폴더 구조 개선
