+++ src/app/(route)/public-data/_components/PUBLIC_DATA_CONST.ts
@@ -7,3 +7,8 @@ export const PUBLIC_DEFAULT_FILTERS = {
   publicRegion: "",
   publicCategory: "",
 };
+
+export const SEARCH_PATH = {
+  lost: "/public-data/lost/search",
+  found: "/public-data/found/search",
+} as const;
