+++ src/app/(route)/public-data/_components/PublicDetailHeader/PublicDetailHeader.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import { useRouter, useSearchParams } from "next/navigation";
 import { DetailHeader } from "@/components/layout";
 import { HeaderSearch } from "@/components/layout/DetailHeader/DetailHeaderParts";
-import { SEARCH_PATH } from "../../_constants/PUBLIC_SEARCH_PATH";
+import { SEARCH_PATH } from "../PUBLIC_DATA_CONST";
 
 const PublicDetailHeader = () => {
   const router = useRouter();
