--- src/app/(route)/public-data/_constants/PUBLIC_SEARCH_PATH.ts
@@ -1,4 +1,0 @@
-export const SEARCH_PATH = {
-  lost: "/public-data/lost/search",
-  found: "/public-data/found/search",
-} as const;
